// sapui5-rt-1.108.31/resources/~20240508083600~/sap/fe/core/converters/ManifestSettings.ts
export enum ActionType {
	DataFieldForAction = "ForAction",
	DataFieldForIntentBasedNavigation = "ForNavigation",
	Default = "Default",
	Primary = "Primary",
	Secondary = "Secondary",
	SwitchToActiveObject = "SwitchToActiveObject",
	SwitchToDraftObject = "SwitchToDraftObject",
	DraftActions = "DraftActions",
	CollaborationAvatars = "CollaborationAvatars",
	DefaultApply = "DefaultApply",
	Menu = "Menu",
	ShowFormDetails = "ShowFormDetails"
}
// ...
type ManifestFormElement = Positionable & {
	type: FormElementType;
	template: string;
	label?: string;
	formatOptions?: FormatOptionsType;
};

export type FormManifestConfiguration = {
	fields: ConfigurableRecord<ManifestFormElement>;
};
// ...
export type FormatOptionsType = {
	hasDraftIndicator?: boolean;
	semantickeys?: string[];
	textLinesEdit?: number;
	textMaxCharactersDisplay?: number;
	textExpandBehaviorDisplay?: string;
};